requirements.txt
gevent
Flask
PyYAML
requests
